undo the move of getStorefrontHeaders and just have the method locally within createHydrogenContext
