+++ packages/hydrogen/src/routing/storefrontRedirect.example.ts
@@ -1,10 +1,9 @@
-import {
-  storefrontRedirect,
-  createStorefrontClient,
-  getStorefrontHeaders,
-} from '@shopify/hydrogen';
+import {storefrontRedirect, createStorefrontClient} from '@shopify/hydrogen';
 import * as remixBuild from '@remix-run/dev/server-build';
-import {createRequestHandler} from '@shopify/remix-oxygen';
+import {
+  createRequestHandler,
+  getStorefrontHeaders,
+} from '@shopify/remix-oxygen';
 
 export default {
   async fetch(request: Request, env: Env, executionContext: ExecutionContext) {
