+++ packages/hydrogen/src/storefrontClient.example.ts
@@ -1,6 +1,9 @@
-import {createStorefrontClient, getStorefrontHeaders} from '@shopify/hydrogen';
+import {createStorefrontClient} from '@shopify/hydrogen';
 import * as remixBuild from '@remix-run/dev/server-build';
-import {createRequestHandler} from '@shopify/remix-oxygen';
+import {
+  createRequestHandler,
+  getStorefrontHeaders,
+} from '@shopify/remix-oxygen';
 
 export default {
   async fetch(
