+++ packages/hydrogen/src/types.d.ts
@@ -51,6 +51,17 @@ export interface HydrogenEnv {
   PUBLIC_CHECKOUT_DOMAIN: string;
 }
 
+export type StorefrontHeaders = {
+  /** A unique ID that correlates all sub-requests together. */
+  requestGroupId: string | null;
+  /** The IP address of the client. */
+  buyerIp: string | null;
+  /** The cookie header from the client  */
+  cookie: string | null;
+  /** The purpose header value for debugging */
+  purpose: string | null;
+};
+
 declare global {
   interface Window {
     privacyBanner: PrivacyBanner;
